Started implementing type checking for branching
Created variable branched_stacks which is a list of TypeStack objects.
It's purpose is that every time a branching block starts a copy of the
stack is pushed to the list. Then when we encounter block's ending word
like endif or done we should check if the stack state is the same
whether or not the block's condition matches.