Fixed type checking for stack altering IF blocks
Before we could not alter the stack in the IF block.
This was a hinderance because we should be able to for example push a
extra value to the stack during IF statement.
To type check the stack altering IF blocks every part of the IF block
MUST have the same stack state after it and there MUST be an ELSE to
catch errors.