Redesigned Lists to use Arrays in the backend
Array are pretty much like Lists before without dynamic resizing.
List objects are an abstraction layer for the arrays and provide easier
management of the capacity of the current List.
Whenever a List requires more space, an array with double the capacity
of the original array will be allocated, List's array pointer will be
updated and the old array will be deallocated.