feat: Implement recursion with true local variables
The return stack contains execution context for each function.
The execution context stores the return address pointer in the `r14` register
and each local variable is stored as an offset from the that register.

When a function is called, a new execution context is created above the
memory address space of the parent function.

Before a function returns, the program switches to the parent's execution context.